bft: check dynamic devices for tftp_server post boot
We don't know the IP address for containers before they are booted for
talking to it from the host. So let's scan again if tftp_server is still
none for tftp servers options and get the ip address instead of falling
back to the legacy wan option... then try wan as the last ditch effort.

Signed-off-by: Matthew McClintock <msm-oss@mcclintock.net>