bft: resolve tftp ip addr after spawning if not explicit
It's possible the ip address is not known from the config provided and
it's only know after it's started. So let's try to resolve the ip
address later in that case so we don't get an error.

Might be simplier to always do this, but this is the minimial invasive
change at the moment. All this device options foo should go in it's own
file at some point to refactor it out of bft as it's getting quite messy

Signed-off-by: Matthew McClintock <msm-oss@mcclintock.net>