deploy-boardfarm-script: add script to deploy docker containers on ethernet interface
This script can be run as follows:

$ ./deploy-boardfarm-nodes.sh eth0 101 106

This will deploy a docker container where eth0 has the access to the
internet and the ssh port (22) and http proxy port (8080) are exposed to
the host on 5NNN and 8NNN based on the vlan tags used (101 -> 106 above)

Also eth1 will be mapped to eth.NNN vlan interface within Linux. This
allows you to mux a single ethernet interface (or perhaps bonded)
interface to a switch as an example and turn 1 ethernet interface into
many interfaces on a switch.

Signed-off-by: Matthew McClintock <msm-oss@mcclintock.net>