devices: debian: add ability to reprovision a board
We want to be able to change a single boards provisioning so let's add
the option to copy new files to the tftp server and rebuild the board
specific DHCP config

Also, remove broken direct inovation of this script to deploy a CMTS
node. Will update this later if that feature is required.

Signed-off-by: Matthew McClintock <msm-oss@mcclintock.net>