bft: make dynamic devices if present in config
Now we check the model field of the device class to see if it matches
the type in the config file and create the device dynamically. The
device is passed the config section.

The old debian device needs to be converted to this model at some point
in the future

Signed-off-by: Matthew McClintock <msm-oss@mcclintock.net>