tests: cdrouter: reorder board reset and parsing config from board
When running CDrouter with a CMTS, we need to get the interface ip for
the CM so we can tell CDrouter how to route traffic to that device.
Well, when we reboot it we have to wait until that device comes up.

This reorders things, and adds some extra checks for waiting to get the
correct info from the board without failing with an error

Signed-off-by: Matthew McClintock <msm-oss@mcclintock.net>