bft: move devices with a feature to end of list
We will select these boards last, if no feature option was passed in.
The idea here is that we will "save" boards with features for the tests
designed to run on them. If no other boards are available we will take
on of the feature boards and give the test run a shot

Signed-off-by: Matthew McClintock <msm-oss@mcclintock.net>