tests: screenshot_gui: fix some issues with getting values from config
The device itself will tell us these so let's use those values directly
and try to build up the http proxy info.

Using the VLAN info is still sort of wrong, since that could change but
for now it works. Going forward it will likely need to always be in the
boardfarm config, which is the default anyways

Signed-off-by: Matthew McClintock <msm-oss@mcclintock.net>