bft: add logger meta to log all function calls for tests and devices
This makes new lgos for all function calls for all device and test
classes. It also saves it to the respective device log that's already
present plus there is a new file for each test log as well.

Then all the new bits are also added to all.log as well

Signed-off-by: Matthew McClintock <msm-oss@mcclintock.net>