Update patch set 3
Patch Set 3:

(3 comments)

The base bits for this look good. I would not add iperf anything specific like that (even ps really) in this class. That belongs somewhere else, parsing text for results.

We are also duplicating text process in several places..

I think it's a good idea to hook this in the o_helper() class directly so we can stream the console to this class (and some helpers that look for specific text- like in the analysis dir) and it will process all text in the background and record live results... no changes to tests themselves.

Still think we need a way to log a value real time from a test too...

Would help to have a very simple example test.

Patch-set: 3
Reviewer: Gerrit User 1016147 <1016147@d5d70762-12d0-45a1-890d-524b12d3f735>