Update patch set 1
Patch Set 1:

As discussed,
This is an ISC-DHCP limitation, it accepts an offset only from -11h to +12h. (in seconds)
We'll figure out a way to get it working with latest RFC standards in a different commit.

The default value is UTC (UTC 0).
Shown in timezone variable. (line 52)

Patch-set: 1
Reviewer: Gerrit User 1016357 <1016357@d5d70762-12d0-45a1-890d-524b12d3f735>