devices: __init__: pexpect session helper funct
Tries to create an object from a class. This is related to the classes
in the devices directory, where the CTOR creates a pexpect session.

Takes:
klass:  str, stringified class name (e.g. "DebianBox") this must exists as a class
name:   str, name attr to be assigned to the obj
ipaddr: str/IPAddress, IP adress for the pexpect session to login to
port:   int, port used for the login
color:  font coloUr for the console

return:
the session obj

Signed-off-by: Michele Gualco <mgualco.contractor@libertyglobal.com>
Change-Id: I81aca465aaba0e2b0bb3560a4f0b085683515ffb