devices: common: error handling on EOF exception
if the copy command does not include the echo for the filename,
then the code generates and EOF exception that can be confusing to
the user. Now handles this exception and print an err msg accordingly.
A small unit test module has been added. It tests only the new
behaviour, but in future it could house more unit tests.

Signed-off-by: Michele Gualco <mgualco.contractor@libertyglobal.com>
Change-Id: Id10c14083ad36557992923c3b4a1efb10530e92b
Reviewed-on: https://review.gerrithub.io/c/lgirdk/boardfarm/+/456391
Tested-by: Matthew McClintock <msm-oss@mcclintock.net>
Reviewed-by: Matthew McClintock <msm-oss@mcclintock.net>