Add class for interacting with a boardfarm server
If boardfarm config is a local file, this feature does nothing.

If boardfarm config is a file hosted on a server, this will
test if it is a "boardfarm server". If not, this feature does
nothing. If yes, this feature lets that server know when it
is using a board, and when it is done.

Signed-off-by: Mike Anderson <mbanderson@uwalumni.com>
Change-Id: Ia1127c281ce9119ee8b7d92dbd72f1aede672437