Update patch set 2
Patch Set 2:

If given a boardfarm config that is a local file path, this class does nothing. Calling "checkout" and "checkin" won't do anything, it's harmless.

If given something that starts with "http", it checks to see if it is a legit bardfarm server. If it's not, then again this class does nothing.

If it is a legit boardfarm server, then "checkout" and "checkin" would actually do something. They would post to the server.

Important question: We need a function that is called on all exits, does that exist? We need that function to call the "checkin" function. Otherwise the boardfarm server won't be notified that a station isn't in use.

Patch-set: 2
Reviewer: Gerrit User 1016622 <1016622@d5d70762-12d0-45a1-890d-524b12d3f735>