devices:debian_isc.py : IPv6 address and PD update
IPv6 addresses and networks are derived from ipaddress.IPv6Interface
Host IPv6 addresses which do not require prefix are left as it is.
Updated the Prefix Delegation for open host devices.
Knowns hosts get a fix prefix for pd, this helps in not having a
conflict in ISP router's pd database.

Signed-off-by: ketan.d.luffy <ktewari.contractor@libertyglobal.com>
Change-Id: I23d3e017d3989e4aeac6944fec5dc293c71f7333
Reviewed-on: https://review.gerrithub.io/c/lgirdk/boardfarm/+/458869
Reviewed-by: Matthew McClintock <msm-oss@mcclintock.net>
Tested-by: Matthew McClintock <msm-oss@mcclintock.net>