Update patch set 7
Patch Set 7:

(7 comments)

This needs to be broken into separate gerrit changes...

Also, bft does not need to be so aware of no shell or no console boards as we need to make a new board type that handles it.

I understand there will be some required refactoring of the common code to handle that case though. The set_router_password is a good example. That should be refactored into a function regardless and we can merge that right away.

Patch-set: 7
Reviewer: Gerrit User 1016147 <1016147@d5d70762-12d0-45a1-890d-524b12d3f735>