tests: docsis: add shortname for cm cfg object
This is not ideal, but it's some info about the CM cfg that was used
that is not 1000s of chars long. We need to look at a better summary, or
maybe we will just get used to the md5.

At least they can be compared against runs to see if they were the same.

Signed-off-by: Matthew McClintock <msm-oss@mcclintock.net>
Change-Id: I487a0b13bd9da421ef680c2600277f0834cf0238
Reviewed-on: https://review.gerrithub.io/c/lgirdk/boardfarm/+/462097
Tested-by: Matthew McClintock <matthew@meldnetworks.com>
Reviewed-by: Mike Anderson <mikeanderson@mcclintock.net>
Reviewed-by: Matthew McClintock <matthew@meldnetworks.com>