Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> Is it taking a long time to ssh to some devices?

Yes, not sure if it's network issues in some locations or just ssh taking a while. This is the start of figuring out how to fix it.

I think one easy solution is to connect to devices in parallel.. probably start a thread per device and try to connect.

Patch-set: 1
Reviewer: Gerrit User 1016147 <1016147@d5d70762-12d0-45a1-890d-524b12d3f735>