Update patch set 4
Patch Set 4:

It maybe me being out of touch, but how can this work with qemu? 
If you look at the boardfarm_config_example.json, qemu does not have a cmts object defined (and how to define it? you would have to create a mock_cmts object that simulates the HW)

moreover this function as it stands is not invoked by any test suite (is this intentional?)
Is this test meant to run on the real HW (and not on travici)?

Patch-set: 4
Reviewer: Gerrit User 1016163 <1016163@d5d70762-12d0-45a1-890d-524b12d3f735>