Update patch set 1
Patch Set 1: Code-Review+1

This tentatively looks OK to me, I was failing on tests specified with '-e' if they failed to parse and we would skip tests in testsuites if they failed to parse.

The reasoning was that if the environment was wrong, we should not skip trying to run the other tests.

We probably need better reporting if we "skipped" a test. Does it should in the skipped test results?

Patch-set: 1
Reviewer: Gerrit User 1016147 <1016147@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=+1