First commit of a Boardfarm class
With this commit you can do this for example:

    import boardfarm
    bf = boardfarm.Boardfarm('http://192.0.0.5/boardfarm/api/bf_config')
    bf.list_stations()

This is the begining of creating useful functions.

Due to the addition of now importing other things inside
of boardfarm/__init__.py, I slightly modified setup.py to grab
version without importing boardfarm. This method is used in
other projects as well.

Signed-off-by: Mike Anderson <mbanderson@uwalumni.com>
Change-Id: I4203aa4ba44f9499572448b23cbddd485bc44563