Create whatchanged.py to find changed tests and features needed
As arguments, give this a start and end git hash, and ALL
boardfarm directories to search.

It returns the names of only tests that changed (separated by "-e" for
boardfarm). It also returns features requested from git log messages.

Command-line options include "--debug" and "--help".

Example use, you have to pass in a start and end git hash pointer as well
as a list of boardfarm "overlays":

    $ ./whatchanged.py HEAD~1 HEAD /tmp/boardfarm-one /tmp/boardfarm-two /tmp/boardfarm
    -e PingTest -e WebTest

Signed-off-by: Mike Anderson <mbanderson@uwalumni.com>
Change-Id: Ic6396a9e8cda887aeeeedfbb8f7af5c7b7fc6f16
Reviewed-on: https://cicd.lgirdk.com/gerrit/c/github/boardfarm/+/421
Tested-by: Jenkins Automation <jenkins@libertyglobal.com>
Reviewed-by: Matthew McClintock <matthew@meldnetworks.com>