devices: add notion of device 'mgr'
Instead of import devices as objects from devices module, you would just
import the mgr and ask it for devices so it can determine the best item
to give you based on location, feature, and any other field we want to
add.

So far this class does not create anything but serves as a repository of
already created devices. Next steps would be to move towards creating
devices when they are requested.

Refactored iperf3 test to use the new API as an example

Change-Id: I53c0bf4d7ea045a76935817a36d2b575e4b639c6
Signed-off-by: Matthew McClintock <msm-oss@mcclintock.net>
Reviewed-on: https://cicd.lgirdk.com/gerrit/c/github/boardfarm/+/726
Tested-by: Jenkins Automation <jenkins@libertyglobal.com>
Reviewed-by: Matthew McClintock <matthew@meldnetworks.com>