make_human_readable: Find html templates in subdirectories
This will find html templates successfully when we adjust them to
be pip installable.

To Do: The html template file to use should probably be a config file
option, or command line argument. Right now it just searchs all the
overlays and uses the first html template file with a specific name
that it finds.

Signed-off-by: Mike Anderson <mbanderson@uwalumni.com>
Change-Id: I338b07ba0800a17ce20add2e929705cd0abde941
Reviewed-on: https://cicd.lgirdk.com/gerrit/c/github/boardfarm/+/1119
Tested-by: Jenkins Automation <jenkins@libertyglobal.com>
Reviewed-by: Mike Anderson <mike@meldnetworks.com>