config: bugfix layerconf imports the correct module
this fixes a bug where if more than 1 layerconf.py is provided
(in diff overlays) the same import layerconf is performed
as it loaded the same module for all imports.
In the layerconfs list (of tuples) the names would be correct but
the associated module objects were the same obj.

Signed-off-by: Michele Gualco <mgualco.contractor@libertyglobal.com>
Change-Id: Id6deeef585197769d52401a528324b6a1a0c0b67
Reviewed-on: https://cicd.lgirdk.com/gerrit/c/github/boardfarm/+/1492
Reviewed-by: Mike Anderson <mike@meldnetworks.com>
Tested-by: Jenkins Automation <jenkins@libertyglobal.com>