devices: remove all placeholder devices
This got a bit more complex but still it's a step closer to easily being
removed in the future. We have a simple place to start emitting a
WARNING message as soon as we are ready

Also one less function to call to setup things for boardfarm as a py3
module.

Change-Id: I7ffce0eb8561966c0e367f44e6bc0731bbe9f79e
Signed-off-by: Matthew McClintock <msm-oss@mcclintock.net>
Reviewed-on: https://cicd.lgirdk.com/gerrit/c/github/boardfarm/+/2113
Tested-by: Jenkins Automation <jenkins@libertyglobal.com>
Reviewed-by: Mike Anderson <mike@meldnetworks.com>