lib: bft pexpect: don't change name once set
This is a work around for a conflicting name space for pexpect and bft.
They both use the name attribute and we need to make sure when it's set
by bft we keep that one and also if pexpect sets it first we can
override it.

This is the complex workaround that can go away once pexpect is not the
same object as a device

Change-Id: I298d2ff53fd7f1111a846ac77d465aacc4d2633a
Signed-off-by: Matthew McClintock <msm-oss@mcclintock.net>
Reviewed-on: https://cicd.lgirdk.com/gerrit/c/github/boardfarm/+/2169
Tested-by: Jenkins Automation <jenkins@libertyglobal.com>
Reviewed-by: Lynn Lin <lynn_lin@compalbn.com>
Reviewed-by: Mike Anderson <mike@meldnetworks.com>