Occasionally you can see the following error when installing dependencies:
ERROR: botocore 1.14.10 has requirement docutils<0.16,>=0.10, but you'll have docutils 0.16 which is incompatible.
This appears to be because both Sphinx and boto3 require it, though Sphinx leads to the installation of a version not supported by Boto3.

From pipdeptree:

boto3==1.11.10
  - botocore [required: >=1.14.10,<1.15.0, installed: 1.14.10]
    - docutils [required: >=0.10,<0.16, installed: 0.15.2]

[...]

Sphinx==2.3.1
[...]
  - docutils [required: >=0.12, installed: 0.15.2]

The proposed solution is to add an explicit requirement in boardfarm's required packages list for docutils.

Change-Id: Iabb841bdb51c1ae20270f0f58d81aa5786b81c9f
Signed-off-by: Filippo Carra <fcarra@libertyglobal.com>
Reviewed-on: https://cicd.lgirdk.com/gerrit/c/github/boardfarm/+/2908
Tested-by: Jenkins Automation <jenkins@libertyglobal.com>
Reviewed-by: Matthew McClintock <matthew@meldnetworks.com>