lib: env_helper: allows for list and non list cmp
When a test environment value is a list and a EnvHelper value is not
checks that the value in the EnvHelper is contained in the test list.
Scenario, a device is configured as IPv4 only, and the test can run
in IPv4, IPv6, and DualStack.
the EnvHelper json would have (this is just a generic example) an entry
like:
    'cfg_mode': 'ipv4'
and the test environment woudl have an entry like:
   'cfg_mode': ['ipv4', 'ipv6', 'dual']
the checker should ensure that the EnvHelper is IN the test
requirements.

Change-Id: I7b67212b10600e60ac283faeabf16963a274a1bd
Signed-off-by: Michele Gualco <mgualco.contractor@libertyglobal.com>
Reviewed-on: https://cicd.lgirdk.com/gerrit/c/github/boardfarm/+/3371
Reviewed-by: Matthew McClintock <matthew@meldnetworks.com>
Tested-by: Jenkins Automation <jenkins@libertyglobal.com>