orchestration, bft_base_test: add teardown feature
bft_base_test will call teardown_class if derived
by a child test.
teardown_class is using TearDown step from orchestration
to call pytest fixture continue_on_fail.
If an action in teardown fails, the error will be stored
as an instance of ContOnFailError.
If the TearDown result has an instance of the fixture
error, test will be marked as FAIL

Signed-off-by: ketan.d.luffy <ktewari.contractor@libertyglobal.com>
Change-Id: I370770ec72ba4cae21d8a203b9682bca74ed55a8
Reviewed-on: https://cicd.lgirdk.com/gerrit/c/github/boardfarm/+/3796
Tested-by: Jenkins Automation <jenkins@libertyglobal.com>
Reviewed-by: Matthew McClintock <matthew@meldnetworks.com>
Reviewed-by: Mike Anderson <mike@meldnetworks.com>