lib: SnmpHelper: add root "oid"s to the dictionary
The callback used to ignore json objects that have the "revisons"
attribute in them. This was forcing root oids like "arris" or "comcast"
to be skipped, since they are defined in the ASN1 module as
objects "oid" and the "revisions" attribute set. We now allow for these
object to be added to the mib dictionary.

Signed-off-by: Michele Gualco <mgualco.contractor@libertyglobal.com>
Change-Id: I4357c65365e5a9c97c4ac56dec2688de63b1f635
Reviewed-on: http://cicd.lgirdk.com/gerrit/c/github/boardfarm/+/8305
Tested-by: Jenkins Automation <jenkins@libertyglobal.com>
(cherry picked from commit 93082155ae226dc3957388b0541b63edf2cc6c35)
Reviewed-on: http://cicd.lgirdk.com/gerrit/c/github/boardfarm/+/8217