devices: softphone: Added an optional argument in answer method
The message "Press a to answer or h to reject call" in the softphone
instance comes only for the first time when connected to the instance.
In scenarios where the Ringing(Incoming Call) has already been checked,
we won't see the answer message in the PJSIP instance. The call can
still be answered by pressing 'a'. Hence adding the optional argument.

Signed-off-by: Nagaraj MS <nnagaraj.contractor@libertyglobal.com>
Change-Id: Ic5c2f38b4f1dbbbcfe0685dd06fe784350a86342
Reviewed-on: http://cicd.lgirdk.com/gerrit/c/github/boardfarm/+/8597
Tested-by: Jenkins Automation <jenkins@libertyglobal.com>
Reviewed-by: Kottesvaran Murugan <mkottesvaran.contractor@libertyglobal.com>
Reviewed-by: Nathuriga Sevvel <nsevvel.contractor@libertyglobal.com>
Reviewed-by: Filippo Carra <fcarra@libertyglobal.com>