bft: tries to keep the "deny_conn_to" alive
Terminal servers have a timeout on idle connections. This will schedule
a periodic "touch" on the "deny_conn_to" boards, so the connection is
kept alive for the duration of the bft session.
This "keep_live" is optional.
NOTE: currently the "deny_conn_to" devices are not checked out of the
dashboard.

Signed-off-by: Michele Gualco <mgualco.contractor@libertyglobal.com>
Change-Id: I8745f5262fc874cd3ae9fb58e40eb9ad31147564
Reviewed-on: http://cicd.lgirdk.com/gerrit/c/github/boardfarm/+/10731
Tested-by: Jenkins Automation <jenkins@libertyglobal.com>
Reviewed-by: Ketan Tewari <ktewari.contractor@libertyglobal.com>