build(pyproject.toml): introduce unsatisfiable pexpect version for Windows
As we rely on pexpect.spawn() we do not support the Windows O.S. As this is a Python-only package we cannot define at build time that it cannot be installed on POSIX-only platform (the O.S. tag in the wheel name is meant to be used for compiled extensions). Introducing a dependency that would fail on Windows is the earliest time we could fail, else we would face an execution issue after succesful installation.

Change-Id: Ib7ba8b4b0bf6950d543e6fa672339ff253c5c8bb
Reviewed-on: http://10.64.38.13:8080/gerrit/c/github/boardfarm/+/14699
Tested-by: Jenkins Automation <jenkins@libertyglobal.com>
Reviewed-by: Michele Gualco <mgualco.contractor@libertyglobal.com>
Reviewed-by: Oleksii Popov <opopov.contractor@libertyglobal.com>