fix: stop autocomp running propeties
Just for testing....
In interactive mode, pressing tab on an object  runs the properties
of the object as code (and may mod the device state). One solution
is to disable the autcomp on tab (and losing some functionality,
annoying), another solution is to use this change where the jedi
getattr is less aggressive.

Signed-off-by: Michele Gualco <msalvatoregualco@libertyglobal.com>
Change-Id: I1e494f0a43f0f4e482a9387d2ffa98bf5fc13e17
Reviewed-on: http://10.64.38.13:8080/gerrit/c/github/boardfarm/+/16991
Tested-by: Jenkins Automation <jenkins@libertyglobal.com>
Reviewed-by: Michele Gualco <mgualco.contractor@libertyglobal.com>