fix(use_cases/dhcp.py): update dhcp methods
Adding additional_args as argument for parse_dhcp_trace to lessen the timeout while filering dhcp  packets. get_dhcp_option_details method will throw an Exception like "out is referenced before assignment" because of this condition failure "if value == str(option):" and out variable is never created, that's why out is initialized as None to not get the exception as "out is referenced before assignment".

Signed-off-by: pooja.jatti <pooja.jatti@infosys.com>
Change-Id: Icab39f00c3e621da75350616dd555f5935b82c29
Reviewed-on: http://10.64.38.13:8080/gerrit/c/github/boardfarm/+/17066
Tested-by: Jenkins Automation <jenkins@libertyglobal.com>
Reviewed-by: Ayush Singhal <aysinghal.contractor@libertyglobal.com>
Reviewed-by: Ketan Tewari <ktewari.contractor@libertyglobal.com>